requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
cffi==1.16.0
colorama==0.4.6
coverage==7.4.3
cryptography==42.0.5
Django==4.2.10
django-db-connection-pool==1.2.4
django-pytest==0.2.0
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
exceptiongroup==1.2.0
factory-boy==3.3.0
Faker==24.3.0
flake8==7.0.0
greenlet==3.0.3
inflection==0.5.1
iniconfig==2.0.0
mccabe==0.7.0
mysqlclient==2.2.4
packaging==23.2
pluggy==1.4.0
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.2.0
PyJWT==2.8.0
pytest==8.0.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.2
six==1.16.0
SQLAlchemy==2.0.28
sqlparams==6.0.1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
